+++ packages/eslint-config-vtex/demo/index.ts
@@ -1,0 +1,7 @@
+/* eslint-disable vtex/prefer-enum-initializers */
+/* eslint-disable prettier/prettier */
+
+const enum Status {
+  Open,
+  Close = 2,
+}