+++ packages/eslint-plugin-vtex/index.js
@@ -5,6 +5,6 @@ module.exports = {
   },
   rules: {
     'prefer-early-return': require('./lib/rules/prefer-early-return'),
-    'enforce-explicit-enum-values': require('./lib/rules/enforce-explicit-enum-values'),
+    'prefer-enum-initializers': require('./lib/rules/prefer-enum-initializers'),
   },
 }