+++ packages/eslint-plugin-vtex/lib/configs/recommended.js
@@ -7,6 +7,6 @@ module.exports = {
         maxStatements: 2,
       },
     ],
-    'vtex/enforce-explicit-enum-values': 'error',
+    'vtex/prefer-enum-initializers': 'warn',
   },
 }